lib: zstd: STL-like implementation of mutex
Added zstd as a Zephyr project to lib.

Implemented mutex, timed_mutex, recursive_mutex and
recursive_timed_mutex that implement all the requirements for the Mutex
C++ named requirements.

Mutexes may be constructed on the stack of dynamically using either a
memory slab or the heap depending on the Kconfig configuration.

Experimentally the zstd implementations may be added to the std
namespace so code can me made more portable between operating systems.

Currenty the code is manually formatted as uncrustify and checkpatch are
not playing nicely with C++ code.

Signed-off-by: Carlos Stuart <carlosstuart1970@gmail.com>